done some routes

+++ app/Models/Disease.php
@@ -20,7 +20,7 @@ class Disease extends Model
     ];
 
     protected $casts=[
-        'patient_id'=>'unsignedBigInteger',
+        'patient_id'=>'integer',
         'name'=>'string',
 
     ];
